requirements.txt
langchain==0.1.16
langchain-groq==0.1.0
langchain-astradb==0.2.0
langchain-core>=0.1.52
langchain-community==0.0.32
python-dotenv
pydantic-settings
flask
flask-cors
requests
huggingface-hub
astrapy>=1.1.0,<2.0
numpy